connected component moved to algos
