+++ docbrown/db/src/algorithms/mod.rs
@@ -4,3 +4,4 @@ pub mod global_triangle_count;
 pub mod local_clustering_coefficient;
 pub mod local_triangle_count;
 pub mod reciprocity;
+pub mod connected_components;
